requirements.txt
streamlit>=1.20.0
beautifulsoup4>=4.11.1
python-Levenshtein>=0.20.9
tldextract>=3.4.0
reportlab>=3.6.12
pandas>=1.5.3
streamlit-card>=1.0.0
streamlit-extras>=0.2.7
